+++ airtel-core-connector/CHANGELOG.md
@@ -2,6 +2,41 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.1.7](https://github.com/mojaloop/airtel-core-connector/compare/v1.1.6...v1.1.7) (2024-08-30)
+
+### [1.1.6](https://github.com/mojaloop/airtel-core-connector/compare/v1.1.5...v1.1.6) (2024-08-30)
+
+### [1.1.5](https://github.com/mojaloop/airtel-core-connector/compare/v1.1.4...v1.1.5) (2024-08-30)
+
+
+### Bug Fixes
+
+* put default values if return value from mojaloop connector is undefined ([803c705](https://github.com/mojaloop/airtel-core-connector/commit/803c7052ac6ed446aecdcbacf62d44b3c055afcc))
+
+### [1.1.4](https://github.com/mojaloop/airtel-core-connector/compare/v1.1.3...v1.1.4) (2024-08-29)
+
+### [1.1.3](https://github.com/mojaloop/airtel-core-connector/compare/v1.1.2...v1.1.3) (2024-08-29)
+
+### [1.1.2](https://github.com/mojaloop/airtel-core-connector/compare/v1.1.1...v1.1.2) (2024-08-29)
+
+### [1.1.1](https://github.com/mojaloop/airtel-core-connector/compare/v1.1.0...v1.1.1) (2024-08-28)
+
+## [1.1.0](https://github.com/mojaloop/airtel-core-connector/compare/v1.0.3...v1.1.0) (2024-08-28)
+
+
+### Features
+
+* implemented get transaction enquirey ([a043d6f](https://github.com/mojaloop/airtel-core-connector/commit/a043d6f1d88c5871da7338e030063f2e59a88cb6))
+
+
+### Bug Fixes
+
+* added changes from update-payer send money ([a5cc399](https://github.com/mojaloop/airtel-core-connector/commit/a5cc3991164d491edb9c5d1f6f3c4243c175b594))
+
+### [1.0.3](https://github.com/mojaloop/airtel-core-connector/compare/v1.0.2...v1.0.3) (2024-08-28)
+
+### [1.0.2](https://github.com/mojaloop/airtel-core-connector/compare/v1.0.1...v1.0.2) (2024-08-26)
+
 ### [1.0.1](https://github.com/mojaloop/airtel-core-connector/compare/v1.3.4...v1.0.1) (2024-08-25)
 
 ### [1.3.4](https://github.com/mojaloop/airtel-core-connector/compare/v1.3.3...v1.3.4) (2024-08-25)